+++ profiles/profile.json
@@ -1,0 +1,3 @@
+{
+    "example": "profile"
+}
